Add option for disabling non-critical warnings
There're tons of 'unused parameter' and 'return type' warnings, because
many methods in BoardGalgo class are WIP. It makes difficult to see
more important results of build without disabling them.

So far, it was 'hard-coded' into basic compilation flags.

It seems to be more elegant solution to have an option, which can be
turned on or off.